Introduce react-router + root/cover/resume routes
This commit introduces react-router and adds three routes:

* `RootRoute`: The default "index" page
* `CoverLetterRoute`: Matt's cover letter
* `ResumeRoute`: Matt's resume

Due to an issue with the `Previewer` component from paged.js, a "hack"
is put in place in `RootRoute` to detect if the user navigated back from
the `ResumeRoute`, and if so, the page is automatically refreshed.
Without this unfortunate hack, the contents of `ResumeRoute` are still
rendered as if the browser's BACK button was never pressed.

As of this commit, the `CoverLetterRoute` component is not implemented.