Add & implement <Loader> component
This commit adds a new component, `<Loader>`, which wraps around a
component that depends on external resume data. The component is a
variant of the code that was in `<Resume>` and `<CoverLetter>` which
handled displaying a "fallback" if the required resume data had not
been loaded. The new `<Loader>` includes a pure CSS "spinner" instead
of the basic "fallback" text. As part of this commit, the existing
`<Resume>` and `<CoverLetter>` components are modified to use the new
loader.